Move ASN1_Time to its own header to avoid <chrono> inclusions
