+++ src/lib/asn1/asn1_print.cpp
@@ -6,6 +6,7 @@
 
 #include <botan/asn1_print.h>
 
+#include <botan/asn1_time.h>
 #include <botan/ber_dec.h>
 #include <botan/bigint.h>
 #include <botan/der_enc.h>
