+++ src/lib/ffi/ffi_cert.cpp
@@ -11,6 +11,7 @@
 #include <memory>
 
 #if defined(BOTAN_HAS_X509_CERTIFICATES)
+   #include <botan/asn1_time.h>
    #include <botan/data_src.h>
    #include <botan/x509_crl.h>
    #include <botan/x509cert.h>
