+++ src/lib/x509/crl_ent.cpp
@@ -8,6 +8,7 @@
 #include <botan/x509_crl.h>
 
 #include <botan/asn1_obj.h>
+#include <botan/asn1_time.h>
 #include <botan/ber_dec.h>
 #include <botan/bigint.h>
 #include <botan/der_enc.h>
