+++ src/lib/x509/x509self.h
@@ -9,6 +9,7 @@
 #define BOTAN_X509_SELF_H_
 
 #include <botan/asn1_obj.h>
+#include <botan/asn1_time.h>
 #include <botan/pkcs10.h>
 #include <botan/pkix_types.h>
 #include <botan/x509cert.h>
